requirements.txt
# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.4.1

# Deep Learning
tensorflow==2.12.0
keras==2.12.0
scikit-learn==1.2.2
numpy==1.23.5
pandas==2.0.1

# API
flask==2.3.2
flask-cors==4.0.0

# Dashboard
streamlit==1.22.0
plotly==5.14.1

# Database
cassandra-driver==3.25.0

# Utilities
python-dotenv==1.0.0 